Update server.js
The images were not being stored in the images folder located in client/public/images. I made the user input work by making the uploaded image be stored first and then uploaded into the database to be displayed. You will also notice that when you run react and node.js the images folder is being ran live as well.